Moved some components into a separate folder
