+++ resources/views/auth/login.blade.php
@@ -7,9 +7,9 @@
     </ul>
 @endif
 
-<x-form action="{{ route('login') }}" method="post">
+<x-form.form action="{{ route('login') }}" method="post">
     <label><input type="text" name="email" id="email"></label>
     <label><input type="password" name="password" id="password"></label>
     <label><input type="checkbox" name="remember" id="remember"> Remember me</label>
     <button type="submit">Log in</button>
-</x-form>
+</x-form.form>
