+++ resources/views/auth/two-factor-auth/challenge.blade.php
@@ -1,4 +1,4 @@
-<x-form action="/two-factor-challenge" method="post">
+<x-form.form action="/two-factor-challenge" method="post">
     <label><input type="text" name="code" id="code"></label>
     <input type="submit" value="Authenticate">
-</x-form>
+</x-form.form>
