--- resources/views/components/delete.blade.php
@@ -1,5 +1,0 @@
-@props(['to'])
-
-<x-form action="{{ $to }}" method="delete">
-    <input type="submit" value="{{ $slot }}">
-</x-form>
