+++ resources/views/components/form/delete.blade.php
@@ -1,0 +1,5 @@
+@props(['to'])
+
+<x-form.form action="{{ $to }}" method="delete">
+    <input type="submit" value="{{ $slot }}">
+</x-form.form>
