+++ resources/views/dashboard.blade.php
@@ -4,6 +4,6 @@ Logged in as: {{ auth()->user()->name }}
     <x-link to="{{ route('fos.index') }}">Admin</x-link>
 </p>
 <x-link to="/two-factor-auth">2FA</x-link>
-<x-form action="/logout" method="post">
+<x-form.form action="/logout" method="post">
     <input type="submit" value="Log out">
-</x-form>
+</x-form.form>
